add RegisterRenderFeaturesCallBack so a video codec can control the render features without having to talk through player
